pyproject.toml
[project]
name = "try-mlx-lm"
version = "0.1.0"
description = "Try MLX-LM package (run local LLMs on Apple Silicon)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "mlx-lm>=0.21.4",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.3.4",
]
